Autorización en Laravel
Autorización en Laravel

+++ app/Http/Controllers/TaskController.php
@@ -56,6 +56,8 @@ class TaskController extends Controller
             return redirect('/tasks');
         }
 
+        $this->authorize('verify', $task);
+
         return view('tasks.edit', ['task' => $task]);
     }
 
@@ -78,6 +80,8 @@ class TaskController extends Controller
             return redirect('/tasks');
         }
 
+        $this->authorize('verify', $task);
+
         $task->title = $request->title;
         $task->save();
         return redirect('/tasks');
@@ -95,6 +99,8 @@ class TaskController extends Controller
         if (empty($task)) {
             return redirect('/tasks');
         }
+
+        $this->authorize('verify', $task);
 
         $task->delete();
         return redirect('/tasks');
